Remove mongo specific implementation of reference sets
